Make generated message structs private
This fixes an issue where messages for a private actor would generate compiler errors due to having a private type in a public interface (i.e. the private actor in the public message type's interface). The message types are entirely internal implementation details and are never accessed outside the module that define them, so they can safely be made private.